simplifying the camera class

+++ NeonEngine/src/graphics/buffers/indexBuffer.cpp
@@ -2,21 +2,21 @@
 
 using namespace neon;
 
-IndexBuffer::IndexBuffer(GLushort* data, GLsizei count)
+IndexBuffer::IndexBuffer(GLuint* data, GLsizei count)
 	: m_count(count) {
 	// Generate a new buffer
 	glGenBuffers(1, &m_ibo);
 	bind();
-	glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(GLushort), data, GL_STATIC_DRAW);
+	glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(GLuint), data, GL_STATIC_DRAW);
 	unbind();
 }
 
-IndexBuffer::IndexBuffer(std::vector<GLushort> &data) {
+IndexBuffer::IndexBuffer(std::vector<GLuint> &data) {
 	// Generate a new buffer
 	m_count = data.size();
 	glGenBuffers(1, &m_ibo);
 	bind();
-	glBufferData(GL_ELEMENT_ARRAY_BUFFER, data.size() * sizeof(GLushort), &data.front(), GL_STATIC_DRAW);
+	glBufferData(GL_ELEMENT_ARRAY_BUFFER, data.size() * sizeof(GLuint), &data.front(), GL_STATIC_DRAW);
 	unbind();
 }
